+++ src/main/java/edu/metrostate/Controller/GroceryListController.java
@@ -1,4 +1,32 @@
+package edu.metrostate.Controller;
+
+import edu.metrostate.Model.Categories.Ingredient;
+
 public class GroceryListController {
+
+    public GroceryListController() {}
+
+    public int addItemToCart(Ingredient ingredient) {
+        return 0;
+    }
+
+    public int incrementQuantity(Ingredient ingredient, int quantity) {
+        return 0;
+    }
+
+    public int removeItemFromCart(Ingredient ingredient) {
+        return 0;
+    }
+
+    public int decrementQuantity(Ingredient ingredient, int quantity) {
+        return 0;
+    }
+
+    public boolean updateTotalItems() {
+        return false;
+    }
+
+
 
 
 }